Add files via upload
This function takes as input a dictionary representing a DAG (directed acyclic graph), such that each node is represented by a key. It returns a nested dictionary called LCA such that the keys of LCA are the keys of p, the keys of LCA[u] are also the keys of p, and the value of LCA[u][v] is a set containing all LCAs of u and v. 
A node x is called a least common ancestor (LCA) of u and v in a tree or a DAG iff x is a common ancestor of u and v and no other common ancestor of u and v is a descendant of x.