Add files via upload
This function requires the TreeSwift library, TreeSwift is a Python library for parsing, manipulating, and iterating over (rooted) tree structures.

In an edge-weighted rooted tree, a diameter is a longest path between any two of its nodes. The length of a path is defined as the sum of all of its edges. This algorithm computes the diameter of an edge-weighted rooted tree in linear-time.

My algorithm works by picking an arbitrary node in the tree and then summing branch lengths to find the furthest node away from it. Then, it does the same thing but starts at the node we just found. This is a well known graph theory technique that produces the furthest path in a binary tree. Then it finds a list of ancestors for these two nodes. Next, it extends the first list by the reverse of the other list, and removes all non unique nodes. This gives us our longest path, and the only thing left to do is to add up all the branch lengths connecting those nodes, which gives us the tree diameter. The algorithm runs linearly, because there are no recursions or nested loops.